pass1

+++ clean-build.sh
@@ -1,5 +1,28 @@
+#!/bin/bash
+set -e
+
+echo ========== Cleaning Bower ==========
+rm -rf vendor
+bower cache clean
+rc=$?
+if [ $rc != 0 ] ; then
+    echo Failed to clean the bower cache
+    exit $rc
+fi
+
+echo ========== Installing Bower Dependencies ==========
+bower install
+rc=$?
+if [ $rc != 0 ] ; then
+    echo Failed to install bower dependencies
+    exit $rc
+fi
+
+echo ========== Cleaning Node Modules ==========
 rm -rf node_modules
+echo ========== Installing Node Modules ==========
 npm install
+
 rm -rf platforms
 ionic platform add android
 ionic platform add ios
